+++ client/app/comment/comment-list.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 
-import {CommentService} from "./comment.service";
+import {CommentService} from "../services/comment.service";
 import {Comment} from "../models/comment.model";
 
 @Component({
